python/code_challenges/hashtable/challenge01/challenge01.py
class TreeNode:
    def __init__(self, value=0, left= None, right= None):
        """
        Initializes a TreeNode with the given value and optional left and right children.
        """
        self.value = value
        self.left = left
        self.right = right

class Node:
    def __init__(self, data, next= None):
        """
        Initializes a Node with the given data and optional next node.
        """
        self.data = data
        self.next = next

class LinkedList:
    def __init__(self):
        """
        Initializes an empty LinkedList with the head set to None.
        """
        self.head = None

    def append(self, data):
        """
        Adds a new node with the given data to the end of the LinkedList.
        """
        new_node = Node(data)
        if self.head is None:
            self.head = new_node
        else:
            current = self.head
            while current.next:
                current = current.next
            current.next = new_node

    def __iter__(self):
        """
        Iterates over the LinkedList, yielding the data of each node.
        """
        current = self.head
        while current:
            yield current.data
            current = current.next

    def __repr__(self):
        """
        Returns a string representation of the LinkedList.
        """
        return ' -> '.join(map(str, self))
    
class HashTable:
    def __init__(self, capacity=10):
        """
        Initializes a HashTable with the given capacity. Default capacity is 10.
        """
        self.capacity = capacity
        self.table = [None] * capacity

    def _hash(self, key):
        """
        Computes the hash index for the given key.
        """
        return hash(key) % self.capacity

    def put(self, key):
        """
        Inserts the given key into the HashTable. Handles collisions using chaining.
        """
        hashed_index = self._hash(key)
        if self.table[hashed_index] is None:
            self.table[hashed_index] = LinkedList()
        self.table[hashed_index].append(key)

    def find(self, key):
        """
        Checks if the given key is present in the HashTable.
        """
        hashed_index = self._hash(key)
        bucket = self.table[hashed_index]
        if bucket is None:
            return False
        for item in bucket:
            if item == key:
                return True
        return False
    
def tree_to_list(root):
    """
    Converts a binary tree to a list representation where `None` represents missing nodes.
    """
    if not root:
        return []

    result = []
    queue = [root]

    while queue:
        node = queue.pop(0)
        if node:
            result.append(node.value)
            queue.append(node.left)
            queue.append(node.right)
        else:
            result.append(None)

    while result and result[-1] is None:
        result.pop()

    return result
    
def has_pair_with_sum(root, target):
    """
    Determines if there exists a pair of nodes in the binary search tree that sum up to the target value.
    """
    hash_table = HashTable()

    def traverse(node):
        if node is None:
            return False

        if traverse(node.left):
            return True

        required = target - node.value
        if hash_table.find(required):
            return True

        hash_table.put(node.value)
        return traverse(node.right)

    return traverse(root)

# Example Usage
if __name__ == "__main__":
    root = TreeNode(5)
    root.left = TreeNode(3)
    root.right = TreeNode(6)
    root.left.left = TreeNode(2)
    root.left.right = TreeNode(4)
    root.right.right = TreeNode(7)

    target = 9

    print("Tree structure as list:", tree_to_list(root))
    print("Target sum:", target)

    print("Output:", has_pair_with_sum(root, target))